# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Main API",
    "version": "1.0.0",
    "description": "API для работы с пользователями, постами и комментариями"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "сервер"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionId"
      }
    },
    "schemas": {
...
      "ErrorUnauthorized": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Unauthorized"
          },
          "error": {
            "type": "string",
            "example": "Invalid token"
          },
          "statusCode": {
            "type": "integer",
            "example": 401
          }
        }
      },
      "ErrorForbidden": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Forbidden"
          },
          "error": {
            "type": "string",
            "example": "You do not have access"
          },
          "statusCode": {
            "type": "integer",
            "example": 403
          }
        }
      },
      "ErrorValidation": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "statusCode": {
            "type": "integer",
            "example": 400
          }
        }
      },
      "ErrorNotFound": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Not found"
          },
          "error": {
            "type": "string",
            "example": "Resource not found"
          },
          "statusCode": {
            "type": "integer",
            "example": 404
          }
        }
      },
...
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clnjak7xj000008l0a9zq3k4f"
          },
          "title": {
            "type": "string",
            "example": "Как писать хороший API"
          },
          "content": {
            "type": "string",
            "example": "API должен быть простым, предсказуемым и документированным..."
          },
          "authorId": {
            "type": "string",
            "example": "clnja0001user123456789"
          },
          "likesCount": {
            "type": "integer",
            "example": 123
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-10-01T12:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-10-02T14:00:00Z"
          }
        }
      },
      "PostCreate": {
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 5,
            "maxLength": 200,
            "example": "Как писать хороший API"
          },
          "content": {
            "type": "string",
            "minLength": 10,
            "example": "API должен быть простым, предсказуемым и документированным..."
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "example": [
              "api",
              "swagger"
            ]
          }
        }
      },
      "PostUpdate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 5,
            "maxLength": 200
          },
          "content": {
            "type": "string",
            "minLength": 10
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 10
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/posts": {
      "post": {
        "summary": "Создать новый пост",
        "description": "Авторизованный пользователь создаёт новый пост с заголовком, контентом и тегами",
        "tags": [
          "Posts"
        ],
        "operationId": "createPost",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Данные нового поста",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пост успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации входных данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorValidation"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизованный доступ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUnauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/search": {
      "get": {
        "summary": "Поиск постов по тексту",
        "description": "Выполняет поиск постов по заданному запросу",
        "tags": [
          "Posts"
        ],
        "operationId": "searchPosts",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "технологии",
            "description": "Поисковый запрос"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Максимальное количество результатов"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Смещение от начала выборки"
          }
        ],
        "responses": {
          "200": {
            "description": "Результаты поиска",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Не указан поисковой запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorValidation"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/top": {
      "get": {
        "summary": "Получить топ постов",
        "description": "Возвращает список наиболее популярных (топовых) постов",
        "tags": [
          "Posts"
        ],
        "operationId": "getTopPosts",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Количество возвращаемых постов"
          }
        ],
        "responses": {
          "200": {
            "description": "Список топ постов",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{id}/like": {
      "post": {
        "summary": "Поставить лайк посту",
        "description": "Авторизованный пользователь ставит лайк указанному посту",
        "tags": [
          "Posts"
        ],
        "operationId": "likePost",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Идентификатор поста, которому ставится лайк"
          }
        ],
        "responses": {
          "204": {
            "description": "Лайк успешно поставлен"
          },
          "400": {
            "description": "Пользователь уже поставил лайк",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorForbidden"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизованный доступ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUnauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{id}": {
      "get": {
        "summary": "Получить пост по ID",
        "description": "Возвращает данные конкретного поста по его идентификатору",
        "tags": [
          "Posts"
        ],
        "operationId": "getPostById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "clnjak7xj000008l0a9zq3k4f",
            "description": "Идентификатор поста"
          }
        ],
        "responses": {
          "200": {
            "description": "Данные поста",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "404": {
            "description": "Пост не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorNotFound"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Обновить пост",
        "description": "Обновляет данные существующего поста, если у пользователя есть права на редактирование",
        "tags": [
          "Posts"
        ],
        "operationId": "updatePost",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Идентификатор поста"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Новые данные поста",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Пост успешно обновлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации входных данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorValidation"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизованный доступ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUnauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Нет прав на редактирование",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorForbidden"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Удалить пост",
        "description": "Удаляет пост по его ID, если у пользователя есть права на удаление",
        "tags": [
          "Posts"
        ],
        "operationId": "deletePost",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Идентификатор поста"
          }
        ],
        "responses": {
          "204": {
            "description": "Пост успешно удален",
            "content": {}
          },
          "401": {
            "description": "Неавторизованный доступ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUnauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Нет прав на удаление",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorForbidden"
                }
              }
            }
          },
          "404": {
            "description": "Пост не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorNotFound"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user/{userId}/posts": {
      "get": {
        "summary": "Получить список постов пользователя с пагинацией",
        "description": "Возвращает список постов указанного пользователя с поддержкой пагинации",
        "tags": [
          "Posts"
        ],
        "operationId": "getUserPosts",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "Идентификатор пользователя (cuid)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Номер страницы",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Количество записей на странице",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список постов пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Пользователь или посты не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorNotFound"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}